[FIX] 디자인 수정
